tests: Make wheel and wheel_contents fixtures global
They can be used by other tests.

Fixes: https://github.com/stanislavlevin/pyproject_installer/issues/21
